This commit makes sets more flexible by requiring you to inject the EntitySystem into your systems then call system.getSet directly. I'm likely to change this to sytem.getNodes shortly, as the notion of an 'entity set' is not right, since each node does not need to contain an entity. However, this mechanism will allow multiple EntitySets to be maintained and referenced by systems as required.
This also removes the concept of a Node as an interface and replaces it with a light duck-type check when an EntitySet is created. Nodes are required to have a prev and next public variable that has the same type as the node itself. This removes a vast number of implicit casts each time a system runs; now only the head node needs to be cast to the correct node class (and since the system must be aware of the node class to function, this is not a bad dependency) rather than for each node. This should offer a significant performance improvement.