Implemented component and entity encoding by encoding to native objects only. This allows simple serialization to JSON, AMF or XML. I haven't yet implemented decoding back to typed objects, so tests comprise checking equivalency of the output objects.
Added encoding and decoding for components, entities and systems

It is unclear whether this is useful... it is a first thought towards allowing [Ember(required)] and [Ember(optional)] tags in Node structures, such that the required elements define the node, but the optional elements are automagically added and removed as they are available...

Interim commit - refactoring to enable [Ember(optional)] tag in Nodes definition.

Completed implementation of [Ember(required)] and [Ember(optional) metadata tags for describing node definitions.