Refactors inspector library.
Makes the previous 'PropertyInspector' an internal implementation detail of the 'ComponentInspector', since all it really does is add a label to the property. This class is renamed to 'PropertyWrapper'.

Gives the previous 'PropertyInput' the status they deserve as 'PropertyInspector'.

Removes tab-handling since the built-in tab-handling does it just as well (probably better).

Ensures that properties in components are consistently positioned by sorting the property labels.

Makes the PropertyInspectorFactory customisable so that custom inspectors can be added as required.