EntityInspectors manage the adding and removal of components onto the entity they map.
This commit also implements an added signal in Entities and exposes some signals in the Ember class in order to allow the Inspector codebase to hook better into system functionality. Ideally this is internal API shared across packages; I'm not sure what the best approach is to stop this from being part of the public API, or whether that's necessary/desirable.